feat(core)!: remove the 25% auto-budget cap; extract domain/cache_budget.rs
The 25% cap was framed as a safety margin, but on the machine that
motivated this whole refactor (128 GiB RAM, 27 GiB weights, ~9 GiB KV) it
discarded ~20 GiB of headroom the subtraction had already proven safe —
the auto-sized cache maxed out at 32 GiB when 76 GiB was actually
available, capping the very feature this work exists to improve. Budget
is now simply total − model − kv − headroom, floored at 1 GiB, no
fractional ceiling.

Also moves compute_auto_cache_ram_mb and its constants out of
server_config.rs into their own domain/cache_budget.rs, alongside the
rest of gglib-core's pure calculations (re-exported from both
crate::domain and crate::server_config so no caller's import path
breaks).
